% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/title-icons.R
\name{title_icons}
\alias{title_icons}
\alias{title_icons_fa5}
\title{Generate text for title icons}
\usage{
title_icons(twitter = NULL, github = NULL)

title_icons_fa5(twitter = NULL, github = NULL)
}
\arguments{
\item{twitter}{Screen name}

\item{github}{User name}
}
\value{
String of html code.
}
\description{
Outputs Twitter and Github icons for Xaringan title slide
}
